+++ tests/modules/postcss-font-family.js
@@ -1,4 +1,4 @@
-module.exports.name = 'cssnano/postcss-fontfamily';
+module.exports.name = 'cssnano/postcss-minify-font-values';
 module.exports.tests = [{
     message: 'should unquote font names',
     fixture: 'h1{font-family:"Helvetica Neue"}',